// src/app/models/caixa_model.ts
export default interface CaixaModel {
  CAI_CODIGO: number;
  CAI_HORAI: string;
  CAI_HORAF?: any;
  CAI_DATAI: string;
  CAI_DATAF?: any;
  CAI_VALORI: number;
  CAI_FUN: number;
  CAI_VALORD: number;
  CAI_VALORC: number;
  CAI_SOBRA?: any;
  CAI_FALTA?: any;
  CAI_BAL: number;
  CAI_LUCRO?: any;
  CAI_PDV: number;
  CAI_VALOR_GAVETA?: any;
  CAI_VALOR_CARTAO: number;
}